requirements.txt
fastapi
uvicorn
python-dotenv
pillow
opencv-python
numpy
pandas
streamlit
google-generativeai
